+++ clients/admin-ui/src/types/api/models/DataUse.ts
@@ -17,6 +17,7 @@ export type DataUse = {
    * Defines the Organization that this resource belongs to.
    */
   organization_fides_key?: string;
+  tags?: Array<string>;
   /**
    * Human-Readable name for this resource.
    */