+++ src/fidesctl/api/routes/util.py
@@ -4,11 +4,11 @@ from typing import Any, Callable
 
 from fastapi import HTTPException, status
 
-from fidesapi.utils.api_router import APIRouter
+from fidesctl.api.utils.api_router import APIRouter
 from fidesctl.core.utils import API_PREFIX as _API_PREFIX
 
 API_PREFIX = _API_PREFIX
-WEBAPP_DIRECTORY = Path("src/fidesapi/build/static")
+WEBAPP_DIRECTORY = Path("src/fidesctl/api/build/static")
 WEBAPP_INDEX = WEBAPP_DIRECTORY / "index.html"